Fix: circular import error
